+++ packages/smart-accounts-kit/src/DelegationFramework/DelegationManager/index.ts
@@ -1,7 +1,8 @@
 import * as constants from './constants';
+import * as decode from './decode';
 import * as encode from './encode';
 import * as execute from './execute';
 import * as read from './read';
 import * as simulate from './simulate';
 
-export { encode, execute, read, simulate, constants };
+export { decode, encode, execute, read, simulate, constants };